refactor(api): move openid-client mock in tests/tooling/mocks
