+++ api/tests/identity-access-management/unit/domain/services/pole-emploi-oidc-authentication-service.test.js
@@ -2,7 +2,7 @@ import { AuthenticationMethod } from '../../../../../src/identity-access-managem
 import { PoleEmploiOidcAuthenticationService } from '../../../../../src/identity-access-management/domain/services/pole-emploi-oidc-authentication-service.js';
 import { config as settings } from '../../../../../src/shared/config.js';
 import { expect, sinon } from '../../../../test-helper.js';
-import { createOpenIdClientMock } from '../../../../tooling/openid-client/openid-client-mocks.js';
+import { createOpenIdClientMock } from '../../../../tooling/mocks/openid-client.mock.js';
 
 describe('Unit | Identity Access Management | Domain | Services | pole-emploi-oidc-authentication-service', function () {
   let openidClient;
